changed custom view, tidied up code
