+++ app/src/main/java/com/example/receipttracking/activities/LoginActivity.kt
@@ -40,6 +40,8 @@ class LoginActivity : AppCompatActivity() {
 
         setContentView(R.layout.activity_login)
 
+        //Delays animations and login to represent loading
+
         object : CountDownTimer(1000, 1000){
             override fun onFinish() {
                 progress_bar_loading.visibility = View.GONE
@@ -53,11 +55,13 @@ class LoginActivity : AppCompatActivity() {
 
 
         button_login.setOnClickListener {
+            //Starts activity with intent
             val intent = Intent(this, ListActivity::class.java)
             startActivity(intent)
         }
 
         textView_skip.setOnClickListener {
+            //Starts activity with intent
             val intent = Intent(this, ListActivity::class.java)
             startActivity(intent)
         }
@@ -73,18 +77,22 @@ class LoginActivity : AppCompatActivity() {
 
         }.setListener(object : Animator.AnimatorListener {
             override fun onAnimationCancel(p0: Animator?) {
+                //Action on animation cancel
 
             }
 
             override fun onAnimationStart(p0: Animator?) {
+                //Action on animation start
 
             }
 
             override fun onAnimationRepeat(p0: Animator?) {
+                //Action on animation repeat
 
             }
 
             override fun onAnimationEnd(p0: Animator?) {
+                //On animation end makes visible
                 animationView_after.visibility = VISIBLE
             }
 
@@ -94,9 +102,6 @@ class LoginActivity : AppCompatActivity() {
 
 
         }
-
-
-
